refactor(types): consolidate callable helpers
