+++ types/instance-hooks.d.ts
@@ -39,8 +39,7 @@ import { FastifyBaseLogger } from './logger'
 import { RouteGenericInterface } from './route'
 import { FastifySchema } from './schema'
 import { FastifyTypeProvider, FastifyTypeProviderDefault } from './type-provider'
-import { AnyFunction, AsyncFunction } from './type-helpers'
-import { ContextConfigDefault, RawReplyDefaultExpression, RawRequestDefaultExpression, RawServerBase, RawServerDefault } from './utils'
+import { AnyFunction, AsyncFunction, ContextConfigDefault, RawReplyDefaultExpression, RawRequestDefaultExpression, RawServerBase, RawServerDefault } from './utils'
 
 type SelectHookHandler<Fn, CallbackHandler, AsyncHandler> =
   Fn extends unknown ? (Fn extends AsyncFunction ? AsyncHandler : CallbackHandler) : Fn
